+++ glasses/src/main/java/com/anezium/rokidrelay/glasses/Constants.kt
@@ -5,4 +5,6 @@ object Constants {
     const val KEY_COMMAND = "rokid_relay.command"
     const val KEY_MEDIA = "rokid_relay.media"
     const val PROTOCOL_VERSION = 1
+    const val BLE_WAKE_SERVICE_UUID = "8b66f35d-7db2-4b3e-9ed4-5fbc5d6b4f01"
+    const val BLE_WAKE_CHARACTERISTIC_UUID = "8b66f35e-7db2-4b3e-9ed4-5fbc5d6b4f01"
 }